implemented triad closure

+++ Rust/chronofold/src/models/vacuum.rs
@@ -3,35 +3,33 @@ use std::collections::HashSet;
 use crate::models::handshake::Handshake;
 use crate::models::monad::Monad;
 
-use serde::Serialize;
-
 /// Friction (λ): Passive Fugacity growth per tick.
 /// High values make Monads crave novelty faster.
-const LAMBDA: f64 = 0.10;
+const LAMBDA: f32 = 0.10;
 
 /// Relaxation (β): The satisfying "cool-down" of Fugacity after a handshake.
 /// Higher values make successful connections more "fulfilling."
-const BETA: f64 = 0.15;
+const BETA: f32 = 0.15;
 
 /// Penalty (α): How quickly Affinity (clinging) spikes after a failed signal.
 /// Lower values (like 0.05) make the network more resilient to rejection.
-const ALPHA: f64 = 0.05;
+const ALPHA: f32 = 0.05;
 
 /// Critical Stress (τ_S): The threshold for action.
 /// When internal stress > τ_S, the Monad is forced to signal.
-const TAU_S: f64 = 0.75;
+const TAU_S: f32 = 0.75;
 
 /// Critical Expansion (τ_F): The threshold for growth.
 /// When Fugacity > τ_F, the Monad gains +1 Dimensional Capacity.
-const TAU_F: f64 = 0.60;
+const TAU_F: f32 = 0.60;
 
 /// Critical Contraction (τ_A): The threshold for retreat.
 /// When Affinity > τ_A, the Monad loses -1 Dimensional Capacity.
-const TAU_A: f64 = 0.85;
+const TAU_A: f32 = 0.85;
 
 #[derive(Clone)]
 pub struct Vacuum {
-    pub tick: u64,
+    pub tick: u32,
     pub max_id: u32,
     pub monads: Vec<Monad>,
     pub handshakes: Vec<Handshake>, // Entanglements / Triad Closures
