+++ Rust/chronofold/src/view_models/monad.rs
@@ -3,6 +3,6 @@ use serde::Serialize;
 #[derive(Serialize)]
 pub struct MonadView {
     pub id: u32,
-    pub fugacity: f64,
-    pub affinity: f64,
+    pub fugacity: f32,
+    pub affinity: f32,
 }
